Move library under composer. Make code PSR compatible.
