+++ .gitignore
@@ -1,0 +1,14 @@
+# Mac OS X dumps these all over the place.
+.DS_Store
+
+# Ignore the /vendor/ directory for people using composer
+/vendor
+
+# If the vendor directory isn't being commited the composer.lock file should also be ignored
+composer.lock
+
+# Ignore IDE specific files
+/.phpintel
+
+# Ignore test build reports
+/build
